Refactor CLI into nested structure
