+++ src/guidellm/cli/benchmark/__init__.py
@@ -1,0 +1,26 @@
+"""Benchmark command group."""
+
+from __future__ import annotations
+
+import click
+
+from guidellm.utils.default_group import DefaultGroupHandler
+
+from .from_file import from_file
+from .run import run
+
+__all__ = ["benchmark"]
+
+
+@click.group(
+    help="Run a benchmark or load a previously saved benchmark report.",
+    cls=DefaultGroupHandler,
+    default="run",
+)
+def benchmark():
+    """Benchmark commands for performance testing generative models."""
+
+
+# Register subcommands
+benchmark.add_command(run)
+benchmark.add_command(from_file)
